+++ app/uk/gov/hmrc/brm/models/gro/Status.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.brm.models
+package uk.gov.hmrc.brm.models.gro
 
 import play.api.libs.functional.syntax._
 import play.api.libs.json.Reads._
